+++ src/main/java/de/maxhenkel/corpse/entities/CorpseEntity.java
@@ -81,9 +81,6 @@ public class CorpseEntity extends CorpseBoundingBoxBase {
     @Override
     public void tick() {
         super.tick();
-        age++;
-        setIsSkeleton(age >= Main.SERVER_CONFIG.corpseSkeletonTime.get());
-
         if (!isNoGravity()) {
             double yMotion = 0D;
             Vector3d motion = getDeltaMovement();
@@ -108,6 +105,10 @@ public class CorpseEntity extends CorpseBoundingBoxBase {
         if (level.isClientSide) {
             return;
         }
+
+        age++;
+        setIsSkeleton(age >= Main.SERVER_CONFIG.corpseSkeletonTime.get());
+
         if (Main.SERVER_CONFIG.corpseForceDespawnTime.get() > 0 && age > Main.SERVER_CONFIG.corpseForceDespawnTime.get()) {
             remove();
             return;