+++ changelog.md
@@ -1,1 +1,1 @@
-- Re-added mod integrations
+- Updated Forge version